moved guard and interceptor to data acess folder of shared folder
