+++ angular/employee/src/app/routes/employee.routes.ts
@@ -1,7 +1,7 @@
 import { Routes } from "@angular/router";
 import { EmployeeComponent } from "../employees/ui/employee/employee.component";
-import { authGuard } from "../shared/guards/auth.guard";
 import { enumRoles } from "../shared/constants.model";
+import { authGuard } from "../shared/data-access/guards/auth.guard";
 
 export const employeeRoutes: Routes = [
   {
