+++ angular/employee/src/app/routes/home.routes.ts
@@ -1,7 +1,7 @@
 import { Routes } from "@angular/router";
 import { HomeComponent } from "../home/feature/home/home.component";
-import { authGuard } from "../shared/guards/auth.guard";
 import { enumRoles } from "../shared/constants.model";
+import { authGuard } from "../shared/data-access/guards/auth.guard";
 
 export const homeRoutes: Routes = [
   {
